Support Multiple init daemons and service programs (#1)
* Add test to determine init system for sshd restart

In the install file, I've added a simple test to determine what init
system is currently on the system, and to select the correct restart
command for sshd that is required.

This is tested on the following OSes in EC2:
* Amazon Linux 2017.09 (using upstart)
* Ubuntu 16.04.3 LTS (using systemd)
* SUSE Enterprise Server 12 SP3 (using systemd)
* RHEL 7.3 (using systemd)

* Hush the stderr message when systemd doesn't exist

This is to ensure that the output for a user is the same as it is
when using the standard codebase.

Message is successfully suppressed on Amazon Linux 2017.09

* Add useradd(1) arg to force user group creation

By default on most distros, `USERGROUPS_ENAB` is set to `yes` which
causes `useradd(1)` to create a group for a user that is added.

There is an implicit dependency on line 179 of the `import_users.sh`
script. I've altered the `useradd(1)` argument set to include
`-U/--user-group` to force this on systems where `USERGROUPS_ENAB`
is set to `no`

This is needed to support SUSE Linux Enterprise Server on EC2.