Allow which(1) to exit and the script to continue
We use `which(1)` to determine if a command is available to us. According
to the man page, `which(1)` will return a return code `1` if the command is
not found. Since the script is run with `-e` as an option in the shebang,
this causes the script to exit prematurely.

I've wrapped the `which(1)` calls in `set +e/set -e` wrappers so that the
script continues gracefully.