Adjust which(1) pipelines to return valid exit codes
On systems where `which(1)` doesn't find a given command, it will return exit
code `1`. This will cause the script to abort under the `-e` option in the
shebang. I've changed the pipeline to capture the return value of which on the
other side of a logical OR, which does work correctly, and thus the pipeline
gets a 0 return code (making `-e` happy), and I get the return value which
tells me whether or not the command exists.

Which makes me happy.