Adjust which(1) blocks to account for shortcircuit eval
I forgot about shortcircuit evaluation. So I've added the success code
in the variable before calling `which(1)`. This means that if the command
is found, the 0 retval will get used in the test. The only time that it
would change that to a non-zero status would be if something wasn't found.

Like before, the whole command pipeline will return success, so `-e` stays
happy as a clam.