Include package build scripts to feed into `fpm` to build deb.
This includes the build scripts and documentation on how to
build a .deb for the project.

It also includes a quick blurb on how to build an RPM from a released version
using the spec file available. At some point, we may want to look at the
possibility of a spec file that builds from the current tree (If that's possible)

EDIT: I modified the build command, so that the cron file isn't a config
file. This is important because if someone were to use `apt-get remove` or
`apt erase`, this file would get left, and thus cron would continue to
execute it.